chore: remove extra project detail page js file
