+++ packages/gatsby-theme-project-portal/src/layouts/ProjectDetailPage.tsx
@@ -1,7 +1,9 @@
 import { BackIcon, ProjectDetail, ProjectDetailProps } from "../components"
 import React, { FunctionComponent } from "react"
-import { Link, withPrefix } from "gatsby"
+import { graphql, Link, withPrefix } from "gatsby"
 import { ImageDataLike } from "gatsby-plugin-image"
+
+export { Head } from "../hooks"
 
 // For the ProjectDetailPage we add the slug which isn't on the layout
 interface ProjectDetailPageProps extends ProjectDetailProps {
@@ -70,3 +72,99 @@ export const ProjectDetailPage: FunctionComponent<
     </>
   )
 }
+
+export default ProjectDetailPage
+
+export const query = graphql`
+  query ProjectDetailPageQuery($slug: String!) {
+    ...HeadData
+    ...LayoutData
+    page: project(slug: { eq: $slug }) {
+      title: question
+      description: summary
+    }
+    defaultContactImage: file(
+      name: { eq: "default-contact" }
+      extension: { in: ["png", "jpg", "jpeg"] }
+      # only match files in the "themeImages" sourced directory:
+      sourceInstanceName: { eq: "themeImages" }
+    ) {
+      childImageSharp {
+        gatsbyImageData(
+          width: 100
+          height: 100
+          placeholder: BLURRED
+          layout: FIXED
+        )
+      }
+    }
+    projectPortalConfig {
+      projectInterestLink
+      staticText {
+        mainContactText: main_contact_text {
+          ongoingText
+          completeText
+        }
+      }
+    }
+    project(slug: { eq: $slug }) {
+      question
+      title
+      summary
+      status
+      opportunityCloses
+      startDate
+      endDate
+      lastModified
+      agency
+      topics {
+        ...TopicDetails
+      }
+      deliverable
+      purpose
+      expertise
+      requirement
+      keyDates
+      priorResearch
+      statusOfData
+      fundingInfo
+      openText: emailContent
+      mainContact {
+        name
+        title
+        employer
+        email
+        image {
+          childImageSharp {
+            gatsbyImageData(
+              width: 100
+              height: 100
+              placeholder: BLURRED
+              layout: FIXED
+            )
+          }
+        }
+      }
+      projectTeam {
+        name
+        title
+        employer
+        email
+        image {
+          childImageSharp {
+            gatsbyImageData(
+              width: 100
+              height: 100
+              placeholder: BLURRED
+              layout: FIXED
+            )
+          }
+        }
+      }
+      faq {
+        text
+        title
+      }
+    }
+  }
+`
